# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HR Attendance & Meeting Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meetings
      parameters:
        - name: upcoming
          in: query
          schema:
            type: ["string", "null"]
          description: "Pass 'true' to get only upcoming meetings"
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create a meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a single meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
    patch:
      operationId: updateMeeting
      tags: [meetings]
      summary: Update a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tasks ───────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: assigneeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Announcements ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, email, department, role, status, joinDate, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        email:     { type: string }
        phone:     { type: ["string", "null"] }
        department: { type: string }
        position:  { type: ["string", "null"] }
        role:      { type: string, enum: [ADMIN, EMPLOYEE] }
        status:    { type: string, enum: [ACTIVE, INACTIVE, ON_LEAVE] }
        avatarUrl: { type: ["string", "null"] }
        joinDate:  { type: string }
        salary:    { type: ["number", "null"] }
        createdAt: { type: string }
...
    Meeting:
      type: object
      required: [id, title, startTime, endTime, organizer, createdAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        description: { type: ["string", "null"] }
        startTime:   { type: string }
        endTime:     { type: string }
        location:    { type: ["string", "null"] }
        meetLink:    { type: ["string", "null"] }
        organizer:   { type: string }
        status:      { type: string, enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED] }
        attendees:
          type: array
          items:
            type: string
        source:      { type: ["string", "null"] }
        googleEventId: { type: ["string", "null"] }
        createdAt:   { type: string }
    MeetingInput:
      type: object
      required: [title, startTime, endTime, organizer]
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        startTime:   { type: string }
        endTime:     { type: string }
        location:    { type: string }
        meetLink:    { type: string }
        organizer:   { type: string }
        status:      { type: string, enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED] }
        attendees:
          type: array
          items:
            type: string
    MeetingUpdate:
      type: object
      properties:
        title:       { type: string }
        description: { type: string }
        startTime:   { type: string }
        endTime:     { type: string }
        location:    { type: string }
        meetLink:    { type: string }
        organizer:   { type: string }
        status:      { type: string, enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED] }
        attendees:
          type: array
          items:
            type: string

    # ── Task ─────────────────────────────────────────────────────────────────
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        description: { type: ["string", "null"] }
        status:      { type: string, enum: [TODO, IN_PROGRESS, DONE] }
        priority:    { type: string, enum: [LOW, MEDIUM, HIGH, URGENT] }
        assigneeId:  { type: ["integer", "null"] }
        assignee:
          $ref: "#/components/schemas/Employee"
        dueDate:     { type: ["string", "null"] }
        createdAt:   { type: string }
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        status:      { type: string, enum: [TODO, IN_PROGRESS, DONE] }
        priority:    { type: string, enum: [LOW, MEDIUM, HIGH, URGENT] }
        assigneeId:  { type: integer }
        dueDate:     { type: string }
    TaskUpdate:
      type: object
      properties:
        title:       { type: string }
        description: { type: string }
        status:      { type: string, enum: [TODO, IN_PROGRESS, DONE] }
        priority:    { type: string, enum: [LOW, MEDIUM, HIGH, URGENT] }
        assigneeId:  { type: integer }
        dueDate:     { type: string }

    # ── Announcement ─────────────────────────────────────────────────────────
...